@@ -11,12 +11,14 @@
   ],
   "main": "modules/griddle.jsx.js",
   "peerDependencies": {
-    "react": ">=0.12.0",
+    "react": "0.14.3",
     "underscore": ">= 1.6.0 < 2"
   },
   "devDependencies": {
+    "babel-jest": "^5.2.1",
+    "babel-loader": "^5.3.3",
     "grunt": "^0.4.5",
-    "grunt-6to5": "^3.0.0",
+    "grunt-babel": "^5.0.3",
     "grunt-contrib-clean": "~0.6.0",
     "grunt-contrib-connect": "^0.8.0",
     "grunt-contrib-copy": "~0.7.0",
@@ -25,73 +27,69 @@
     "grunt-jsxhint": "^0.4.0",
     "grunt-markdown": "~0.6.1",
     "grunt-open": "^0.2.3",
-    "grunt-react": "~0.10.0",
-    "grunt-webpack": "~1.0.8",
-    "6to5-jest": "*",
-    "jest-cli": "*",
-    "jsx-loader": "~0.12.x",
+    "grunt-webpack": "^1.0.11",
+    "jasmine-core": "^2.3.4",
+    "karma": "^0.13.15",
+    "karma-babel-preprocessor": "^6.0.1",
+    "karma-chrome-launcher": "^0.2.1",
+    "karma-jasmine": "^0.3.6",
+    "karma-phantomjs-launcher": "^0.2.1",
+    "karma-webpack": "^1.7.0",
+    "phantomjs": "^1.9.18",
+    "phantomjs-polyfill": "0.0.1",
+    "react": "^0.14.3",
+    "react-addons-test-utils": "0.14.3",
     "react-chartist": "~0.2.1",
-    "react-tools": "~0.12.x",
+    "react-dom": "0.14.3",
     "underscore": "~1.6.0",
-    "webpack": "~1.3.3-beta1",
-    "webpack-dev-server": "~1.7.0"
+    "webpack": "1.12.3",
+    "webpack-dev-server": "1.12.1"
   },
   "scripts": {
-    "test": "jest"
+    "test": "karma start --browsers PhantomJS",
+    "test-single": "karma start --singleRun true --browsers PhantomJS"
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/DynamicTyped/Griddle.git"
+    "url": "git://github.com/GriddleGriddle/Griddle.git"
   },
   "author": {
-    "name": "Ryan Lanciaux"
+    "name": "GriddleGriddle"
   },
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/DynamicTyped/Griddle/issues"
+    "url": "https://github.com/GriddleGriddle/Griddle/issues"
   },
-  "jest": {
-    "rootDir": "scripts",
-    "scriptPreprocessor": "../node_modules/6to5-jest",
-    "testFileExtensions": [
-      "es6",
-      "js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "es6"
-    ],
-    "unmockedModulePathPatterns": [
-      "react"
-    ]
-  },
-  "homepage": "https://github.com/DynamicTyped/Griddle",
-  "gitHead": "b99473ce2ba0ccc564ddbcccb402c694bb04be61",
-  "_id": "griddle-react@0.2.14",
-  "_shasum": "e4e823513ab891d89a10ea85089343969c57695a",
-  "_from": "griddle-react@*",
-  "_npmVersion": "2.11.2",
-  "_nodeVersion": "0.10.35",
+  "homepage": "https://github.com/GriddleGriddle/Griddle",
+  "gitHead": "aeeee15853574be17e4fcc995fe94d174e954381",
+  "_id": "griddle-react@0.3.0",
+  "_shasum": "84724a6953e7c19ec3079ed0f822c2718eba33ce",
+  "_from": "griddle-react@>=0.3.0 <0.4.0",
+  "_npmVersion": "2.14.4",
+  "_nodeVersion": "4.1.2",
   "_npmUser": {
     "name": "joellanciaux",
     "email": "joel.lanciaux@gmail.com"
   },
+  "dist": {
+    "shasum": "84724a6953e7c19ec3079ed0f822c2718eba33ce",
+    "tarball": "http://registry.npmjs.org/griddle-react/-/griddle-react-0.3.0.tgz"
+  },
   "maintainers": [
-    {
-      "name": "ryanlanciaux",
-      "email": "ryan.lanciaux@gmail.com"
-    },
     {
       "name": "joellanciaux",
       "email": "joel.lanciaux@gmail.com"
+    },
+    {
+      "name": "kevinold",
+      "email": "kevinold@gmail.com"
+    },
+    {
+      "name": "ryanlanciaux",
+      "email": "ryan.lanciaux@gmail.com"
     }
   ],
-  "dist": {
-    "shasum": "e4e823513ab891d89a10ea85089343969c57695a",
-    "tarball": "http://registry.npmjs.org/griddle-react/-/griddle-react-0.2.14.tgz"
-  },
   "directories": {},
-  "_resolved": "https://registry.npmjs.org/griddle-react/-/griddle-react-0.2.14.tgz",
+  "_resolved": "https://registry.npmjs.org/griddle-react/-/griddle-react-0.3.0.tgz",
   "readme": "ERROR: No README data found!"
 }
