@@ -26,35 +39,48 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "history": "1.12.3",
     "invariant": "^2.0.0",
     "warning": "^2.0.0"
   },
+  "peerDependencies": {
+    "history": "1.13.x"
+  },
   "devDependencies": {
-    "assert": "1.3.0",
     "babel": "^5.4.7",
     "babel-core": "^5.4.7",
     "babel-eslint": "^3.1.23",
     "babel-loader": "^5.0.0",
+    "babel-plugin-dev-expression": "^0.1.0",
     "bundle-loader": "^0.5.2",
+    "codecov.io": "^0.1.6",
+    "coveralls": "^2.11.4",
     "css-loader": "^0.19.0",
-    "eslint": "1.4.0",
-    "eslint-plugin-react": "3.3.2",
-    "expect": "1.10.0",
+    "eslint": "^1.7.3",
+    "eslint-config-rackt": "^1.1.0",
+    "eslint-plugin-react": "^3.6.3",
+    "expect": "^1.12.0",
     "express": "^4.13.3",
     "express-urlrewrite": "^1.2.0",
-    "karma": "^0.13.8",
+    "gzip-size": "^3.0.0",
+    "history": "^1.12.5",
+    "isparta-loader": "^1.0.0",
+    "karma": "^0.13.13",
     "karma-browserstack-launcher": "^0.1.4",
     "karma-chrome-launcher": "^0.2.0",
-    "karma-firefox-launcher": "^0.1.6",
+    "karma-coverage": "^0.5.3",
     "karma-mocha": "^0.2.0",
     "karma-mocha-reporter": "^1.1.1",
     "karma-sourcemap-loader": "^0.3.5",
     "karma-webpack": "^1.7.0",
     "mocha": "^2.0.1",
+    "pretty-bytes": "^2.0.1",
     "qs": "^4.0.0",
-    "react": "0.13.x",
-    "rf-changelog": "^0.4.0",
+    "react": "^0.14.0",
+    "react-addons-css-transition-group": "^0.14.0",
+    "react-addons-test-utils": "0.14.0",
+    "react-dom": "^0.14.0",
+    "react-static-container": "^1.0.0",
+    "rimraf": "^2.4.3",
     "style-loader": "^0.12.4",
     "webpack": "^1.4.13",
     "webpack-dev-middleware": "^1.2.0"
