@@ -34,7 +34,7 @@ module.exports = function(grunt) {
 
     grunt.log.writeln('NODE_ENV: ' + process.env.NODE_ENV);
 
-    var sauceConnect = require('./e2e/sauce_connect');
+    var sauceDns = grunt.option('sauceDns') || '127.0.0.1';
     var sauceJobName = grunt.option('sauceJobName') || 'e2e_' + currentDateTime;
     var sauceKey = grunt.option('sauceKey');
 
@@ -414,34 +414,60 @@ module.exports = function(grunt) {
         },
 
         protractor: {
+            sauce_win7_chrome : {
+                options: {
+                    configFile: './e2e/config/sauce.chrome.win7.protractor.conf.js',
+                    args: {
+                        baseUrl   : baseUrl
+                    }
+                }
+            },
             sauce_linux_chrome : {
                 options: {
                     configFile: './e2e/config/sauce.chrome.linux.protractor.conf.js',
-                    baseUrl   : baseUrl
+                    args: {
+                        baseUrl   : baseUrl
+                    }
                 }
             },
             sauce_linux_firefox : {
                 options: {
                     configFile: './e2e/config/sauce.firefox.linux.protractor.conf.js',
-                    baseUrl   : baseUrl
+                    args: {
+                        baseUrl   : baseUrl
+                    }
                 }
             },
             sauce_multi_browser : {
                 options: {
                     configFile: './e2e/config/sauce.multi.browser.protractor.conf.js',
-                    baseUrl   : baseUrl
+                    args: {
+                        baseUrl   : baseUrl
+                    }
                 }
             },
             local              : {
                 options: {
                     configFile: './e2e/config/local.protractor.conf.js',
-                    baseUrl   : baseUrl
+                    args: {
+                        baseUrl   : baseUrl
+                    }
+                }
+            },
+            local_sauce              : {
+                options: {
+                    configFile: './e2e/config/local.sauce.protractor.conf.js',
+                    args: {
+                        baseUrl   : baseUrl
+                    }
                 }
             },
             local_data_gen : {
                 options: {
                     configFile: './e2e/config/local.dataGen.protractor.conf.js',
-                    baseUrl   : baseUrl
+                    args: {
+                        baseUrl   : baseUrl
+                    }
                 }
             }
         },
@@ -474,9 +500,34 @@ module.exports = function(grunt) {
                 NODE_ENV                    : 'local',
                 NODE_TLS_REJECT_UNAUTHORIZED: 0,
                 ENV_TUNNEL_NAME             : tunnelIdentifier,
+                SAUCE_DNS                   : sauceDns,
                 SAUCE_JOB_NAME              : sauceJobName,
                 SAUCE_KEY                   : sauceKey,
                 DOMAIN                      : baseUrl
+            }
+        },
+
+        sauce_connect: {
+            local: {
+                options: {
+                    username        : 'sbg_qbse',
+                    accessKey       : sauceKey,
+                    proxy           : httpProxy,
+                    tunnelIdentifier: tunnelIdentifier,
+                    verbose         : grunt.option('verbose') === true,
+                    logger          : console.log,
+                    dns             : sauceDns
+                }
+            },
+            aws: {
+                options: {
+                    username        : 'sbg_qbse',
+                    accessKey       : sauceKey,
+                    proxy           : httpProxy,
+                    tunnelIdentifier: tunnelIdentifier,
+                    verbose         : grunt.option('verbose') === true,
+                    logger          : console.log
+                }
             }
         },
 
@@ -652,7 +703,7 @@ module.exports = function(grunt) {
     });
 
     grunt.registerTask('codeStandards', [
-        'lint',
+        'lint'
     ]);
 
     grunt.registerTask('testClientOnly', function() {
@@ -719,24 +770,7 @@ module.exports = function(grunt) {
             ]);
         }
 
-        if (target === 'e2e') {
-            return grunt.task.run([
-                'env:e2e',
-                'sauce_connect',
-                'protractor:sauce_linux_chrome',
-                'sauce-connect-close'
-            ]);
-        }
-
-        if (target === 'e2eMulti') {
-            return grunt.task.run([
-                'env:e2e',
-                'sauce_connect',
-                'protractor:sauce_multi_browser',
-                'sauce-connect-close'
-            ]);
-        }
-
+        // Run your protractor tests locally against your dev env
         if (target === 'e2eLocal') {
             return grunt.task.run([
                 'clean:server',
@@ -745,11 +779,31 @@ module.exports = function(grunt) {
             ]);
         }
 
-        if (target === 'e2eDataGen') {
+        // Run a protractor spec file that will generate you a ticket, realm and app in your local dev
+        if (target === 'e2eLocalDataGen') {
             return grunt.task.run([
                 'clean:server',
                 'autoprefixer',
                 'protractor:local_data_gen'
+            ]);
+        }
+
+        // Run your protractor tests in Sauce Labs against your local dev env
+        if (target === 'e2eLocalSauce') {
+            return grunt.task.run([
+                'env:local',
+                'sauce_connect:local',
+                'protractor:local_sauce',
+                'sauce-connect-close'
+            ]);
+        }
+
+        // Run your protractor tests via Sauce Labs against an existing AWS swimlane
+        if (target === 'e2eAWSSauce') {
+            return grunt.task.run([
+                'env:e2e',
+                'sauce_connect:aws',
+                'protractor:sauce_win7_chrome'
             ]);
         }
 
@@ -765,11 +819,9 @@ module.exports = function(grunt) {
 
     });
 
-
     grunt.registerTask('testIntegration', function() {
         grunt.task.run(['test:integration']);
     });
-
 
     grunt.registerTask('testE2ELocal', function() {
         grunt.task.run(['test:e2eLocal']);
@@ -801,51 +853,13 @@ module.exports = function(grunt) {
         'build'
     ]);
 
-    /* global console:true */
-    grunt.registerTask('sauce_connect', 'Grunt plug-in to download and launch Sauce Labs Sauce Connect', function() {
-        var options = options({
-            username        : 'sbg_qbse',
-            accessKey       : sauceKey,
-            proxy           : httpProxy,
-            tunnelIdentifier: tunnelIdentifier,
-            verbose         : grunt.option('verbose') === true,
-            logger          : console.log
-        });
-
-        var done = this.async();
-
-        var tunnel = {};
-        sauceConnect.setOptions(options);
-
-        grunt.log.writeln('Found tunnelIdentifier: ' + options.tunnelIdentifier);
-
-        if (tunnel.process) {
-            grunt.log.writeln('Existing'.cyan + ' Sauce Connect tunnel: ' + tunnel.tid);
-            sauceConnect.close(sauceConnect.open(done));
-        } else {
-            tunnel = sauceConnect.open(done);
-        }
-    });
-
-    grunt.registerTask('sauce-connect-close', 'Closes the current Sauce Connect tunnel', function() {
-        sauceConnect.close(this.async());
-    });
-
     grunt.registerTask('lint', 'Run eslint on code', function(){
         return grunt.task.run([
             'shell:lint',
         ]);
     });
 
-    grunt.registerTask('makeProdNodeModules', 'Creates a production copy of node_modules folder for zip to nexus', function() {
-        // delete any old modules in dist dir
-        grunt.task.run(['clean:modulesProd']);
-
-        //copy current modules to prod modules dir
-        grunt.task.run(['copy:modulesProd']);
-
-    });
-
     grunt.loadNpmTasks('grunt-shell-spawn');
     grunt.loadNpmTasks('grunt-webpack');
+    grunt.loadNpmTasks('grunt-sauce-connect-launcher');
 };
