@@ -1,78 +1,78 @@
 import React from 'react';
-import {Nav, Tooltip, OverlayTrigger, Glyphicon} from 'react-bootstrap';
+import {Tooltip, OverlayTrigger, Glyphicon} from 'react-bootstrap';
 import {Link} from 'react-router';
 import Loader  from 'react-loader';
-import './leftNav.scss';
 import {I18nMessage} from '../../utils/i18nMessage';
+import qbLogo from '../../assets/images/intuit_logo_white.png';
+import ReactCSSTransitionGroup from 'react/lib/ReactCSSTransitionGroup';
+import GlobalActions from '../global/globalActions';
+import AppsList from './appsList';
+import TablesList from './tablesList';
+import ReportsList from './reportsList';
+import Hicon from '../harmonyIcon/harmonyIcon';
+
+import './leftNav.scss';
 
 let LeftNav = React.createClass({
 
-    getGlyphName(item) {
-
-        if (item.icon) {
-            return item.icon;
-        } else {
-            return 'th-list';
-        }
+    propTypes: {
+        open:React.PropTypes.bool.isRequired,
+        appsListOpen:React.PropTypes.bool.isRequired,
+        selectedAppId:React.PropTypes.string,
+        selectedTableId:React.PropTypes.string,
+        reportsData:React.PropTypes.object,
+        selectedReportId:React.PropTypes.string,
+        showReports:React.PropTypes.bool.isRequired,
+        onToggleAppsList:React.PropTypes.func,
+        onSelectReports:React.PropTypes.func,
+        onSelect:React.PropTypes.func,
+        globalActions:React.PropTypes.array
     },
 
-    buildHeadingItem: function(item, loadingCheck) {
-
-        if (this.props.open) {
-            return (
-                <li key={item.key}>
-                    <Loader scale={.5} right={'90%'} loaded={!loadingCheck}/>
-                    <a className="heading"><I18nMessage message={item.key}/></a>
-                </li>);
-        } else {
-            return (<li key={item.key}><a className="heading"></a></li>);
-        }
-    },
-
-    buildNavItem: function(item) {
-
-        let label = item.name;
-        if (item.key) {
-            label = (<I18nMessage message={item.key}/>);
-        }
-
-        const tooltip = (<Tooltip className={ this.props.open ? 'leftNavTooltip' : 'leftNavTooltip show' }
-                                  id={label}>{label}</Tooltip>);
-
-        let selectedClass = item.id && (item.id.toString() === this.props.reportID) ? 'selected' : '';
-
-        return (
-            <OverlayTrigger key={label} placement="right" overlay={tooltip}>
-                <li className={selectedClass}>
-                    <Link className="leftNavLink" to={item.link}>
-                        <Glyphicon glyph={this.getGlyphName(item)}/> {label}
-                    </Link>
-                </li>
-            </OverlayTrigger>
+    /**
+     * create a branding section (logo with an apps toggle if an app is selected)
+     */
+    createBranding() {
+        let app = _.findWhere(this.props.apps, {id: this.props.selectedAppId});
+        return (this.props.open &&
+            <div className="branding">
+                <img src={qbLogo} />
+                {this.props.selectedAppId &&
+                    <div className="appsToggle" onClick={this.props.toggleAppsList}>{app ? app.name : ''}&nbsp;
+                        <Hicon icon="chevron-down"/>
+                    </div>
+                }
+            </div>
         );
     },
 
-    render: function() {
+    getAppTables(appId) {
+        let app = _.findWhere(this.props.apps, {id: appId});
 
+        return app ? app.tables : [];
+    },
+
+    onSelectApp() {
+        this.props.toggleAppsList(false);
+    },
+    render() {
         return (
 
-            <div className={"leftMenu " + (this.props.open ? "open" : "closed")}>
+            <div className={"leftNav " + (this.props.open ? "open " : "closed ") + (this.props.appsListOpen ? "appsListOpen" : "")}>
+                {this.createBranding()}
 
-                <Nav stacked activeKey={1}>
-                    {this.props.items ? this.props.items.map((item) => {
-                        return item.heading ?
-                            this.buildHeadingItem(item) :
-                            this.buildNavItem(item);
-                    }) : null}
-                    {this.props.reportsData ? this.buildHeadingItem({key: 'nav.reportsHeading'}, this.props.reportsData.loading) : null}
-                    {this.props.reportsData && this.props.reportsData.list ? this.props.reportsData.list.map((item) => {
-                        return this.buildNavItem(item);
-                    }) : ''}
+                <ReactCSSTransitionGroup transitionName="leftNavList" component="div" className={"transitionGroup"} transitionEnterTimeout={300} transitionLeaveTimeout={300}>
+                    {!this.props.selectedAppId || this.props.appsListOpen ?
+                        <AppsList key={"apps"} {...this.props} onSelectApp={this.onSelectApp}  /> :
+                        <TablesList key={"tables"} {...this.props} showReports={(id)=>{this.props.onSelectReports(id);} } getAppTables={this.getAppTables}/> }
 
-                </Nav>
+                </ReactCSSTransitionGroup>
+
+                {this.props.globalActions && <GlobalActions actions={this.props.globalActions} onSelect={this.props.onSelect}/>}
+
+                <ReportsList open={this.props.open} onSelect={this.props.onSelect} reportsOpen={this.props.showReports} onBack={this.props.onHideReports} reportsData={this.props.reportsData} />
 
             </div>
-
         );
     }
 });
