@@ -1,8 +1,11 @@
 
 export const TOGGLE_LEFT_NAV = 'TOGGLE_LEFT_NAV';
-export const SHOW_TROUSER = 'SHOW_TROUSER';
-export const HIDE_TROUSER = 'HIDE_TROUSER';
+export const SHOW_TROWSER = 'SHOW_TROWSER';
+export const HIDE_TROWSER = 'HIDE_TROWSER';
 export const CHANGE_LOCALE = 'CHANGE_LOCALE';
+export const TOGGLE_APPS_LIST = 'TOGGLE_APPS_LIST';
+export const SHOW_REPORTS = 'SHOW_REPORTS';
+export const HIDE_REPORTS = 'HIDE_REPORTS';
 
 export const TOGGLE_SEARCH = 'TOGGLE_SEARCH';
 export const SEARCH_FOR = 'SEARCH_FOR';
@@ -10,7 +13,10 @@ export const SEARCHING = 'SEARCH';
 
 export const SHOW_NEW_ITEMS = 'SHOW_NEW_ITEMS';
 export const LOAD_APPS = 'LOAD_APPS';
-export const LOAD_APPS_WITH_TABLES = 'LOAD_APPS_WITH_TABLES';
+export const LOAD_APPS_SUCCESS = 'LOAD_APPS_SUCCESS';
+export const LOAD_APPS_FAILED = 'LOAD_APPS_FAILED';
+export const SELECT_APP = 'SELECT_APP';
+export const SELECT_TABLE = 'SELECT_TABLE';
 
 export const LOAD_REPORTS = 'LOAD_REPORTS';
 export const LOAD_REPORTS_SUCCESS = 'LOAD_REPORTS_SUCCESS';
