@@ -3,33 +3,37 @@ import * as actions from '../constants/actions';
 import Fluxxor from 'fluxxor';
 import Locale from '../locales/locales';
 import Logger from '../utils/logger';
-var logger = new Logger();
+
+let logger = new Logger();
 
 let NavStore = Fluxxor.createStore({
 
     initialize: function() {
         this.state = {
             leftNavOpen: true,
+            appsListOpen: false,
             searchBarOpen: false,
             searching:false,
-            trouserOpen: false,
+            trowserOpen: false,
             newItemsOpen: false,
+            showReports: false,
             leftNavItems: []
         };
 
         this.setLocaleBundle();
 
         this.bindActions(
-            actions.SHOW_TROUSER, this.onShowTrouser,
-            actions.HIDE_TROUSER, this.onHideTrouser,
+            actions.SHOW_TROWSER, this.onShowTrowser,
+            actions.HIDE_TROWSER, this.onHideTrowser,
+            actions.SHOW_REPORTS, this.onShowReports,
+            actions.HIDE_REPORTS, this.onHideReports,
             actions.SHOW_NEW_ITEMS, this.onShowNewItems,
             actions.TOGGLE_LEFT_NAV, this.onToggleLeftNav,
+            actions.TOGGLE_APPS_LIST, this.onToggleAppsList,
             actions.TOGGLE_SEARCH, this.onToggleSearch,
             actions.SEARCHING, this.onSearching,
             actions.CHANGE_LOCALE, this.onChangeLocale
         );
-
-        this.state.leftNavItems.push({key: 'nav.home', link:'/apps', icon:'home'});
     },
 
     onChangeLocale: function() {
@@ -38,13 +42,21 @@ let NavStore = Fluxxor.createStore({
         this.emit('change');
     },
 
-    onShowTrouser: function() {
-        this.state.trouserOpen = true;
+    onShowTrowser: function() {
+        this.state.trowserOpen = true;
+        this.emit('change');
+    },
+    onHideTrowser: function() {
+        this.state.trowserOpen = false;
         this.emit('change');
     },
 
-    onHideTrouser: function() {
-        this.state.trouserOpen = false;
+    onShowReports: function() {
+        this.state.showReports = true;
+        this.emit('change');
+    },
+    onHideReports: function() {
+        this.state.showReports = false;
         this.emit('change');
     },
     onToggleSearch: function() {
@@ -55,8 +67,29 @@ let NavStore = Fluxxor.createStore({
         this.state.searching = searching;
         this.emit('change');
     },
-    onToggleLeftNav: function() {
-        this.state.leftNavOpen = !this.state.leftNavOpen;
+
+    /*
+     * toggle left nav open/closed
+     * @param show force shown/hidden
+     */
+    onToggleLeftNav: function(show) {
+        if (show === false || show === true) {
+            this.state.leftNavOpen = show;
+        } else {
+            this.state.leftNavOpen = !this.state.leftNavOpen;
+        }
+        this.emit('change');
+    },
+    /*
+     * toggle apps list
+     * @param show force shown/hidden
+     */
+    onToggleAppsList: function(show) {
+        if (show === false || show === true) {
+            this.state.appsListOpen = show;
+        } else {
+            this.state.appsListOpen = !this.state.appsListOpen;
+        }
         this.emit('change');
     },
     onShowNewItems: function() {
