@@ -31,7 +31,7 @@ describe('Report Actions functions -- success', () => {
         spyOn(flux.dispatchBinder, 'dispatch');
         reportActions.__Rewire__('ReportService', mockReportService);
 
-        promise = flux.actions.loadReports(testData);
+        promise = flux.actions.loadReports(testData.appId, testData.tblId);
 
         //  expect a load report event to get fired before the promise returns
         expect(flux.dispatchBinder.dispatch).toHaveBeenCalledWith(actions.LOAD_REPORTS);
