@@ -7,6 +7,7 @@
 // jshint sub: true
 (function() {
     'use strict';
+
     // In order to manage the async nature of Protractor with a non-Angular page use the ExpectedConditions feature
     var EC = protractor.ExpectedConditions;
     //Require the e2e base class and constants modules
@@ -18,8 +19,9 @@
     var requestSessionTicketPage = require('./requestSessionTicket.po.js');
     var requestAppsPage = require('./requestApps.po.js');
     var ReportServicePage = require('./reportService.po.js');
-    describe('Report Griddle Layout Tests', function() {
+    var reportServicePage = new ReportServicePage();
 
+    describe('Report Page Layout Tests', function() {
         var widthTest = 1024;
         var heightTest = 2024;
         e2eBase.setBaseUrl(browser.baseUrl);
@@ -44,7 +46,7 @@
             tableToFieldToFieldTypeMap['table 1']['Url'] = {fieldType: consts.SCALAR, dataType: consts.URL};
             tableToFieldToFieldTypeMap['table 1']['Duration'] = {fieldType: consts.SCALAR, dataType : consts.DURATION};
             tableToFieldToFieldTypeMap['table 1']['Email'] = {fieldType: consts.SCALAR, dataType : consts.EMAIL_ADDRESS};
-            tableToFieldToFieldTypeMap['table 1']['Rating'] = {fieldType: consts.SCALAR, dataType: consts.RATING};
+            tableToFieldToFieldTypeMap['table 1']['Rating 2'] = {fieldType: consts.SCALAR, dataType: consts.RATING};
             tableToFieldToFieldTypeMap['table 2'] = {};
             tableToFieldToFieldTypeMap['table 2']['Text Field'] = {fieldType: consts.SCALAR, dataType: consts.TEXT};
             tableToFieldToFieldTypeMap['table 2']['Rating Field'] = {fieldType: consts.SCALAR, dataType : consts.RATING};
@@ -77,82 +79,120 @@
                         requestAppsPage.get(e2eBase.getRequestAppsPageEndpoint(realmName));
                         // Define the window size
                         e2eBase.resizeBrowser(widthTest, heightTest);
+                        // Wait for the left nav to load
+                        reportServicePage.waitForElement(reportServicePage.appsListDivEl).then(function() {
+                            // Select the app
+                            reportServicePage.appLinksElList.get(0).click();
+                            done();
+                        });
+                    });
+                });
+            });
+        });
+
+        /**
+         * Before each test starts just make sure the table list div has loaded
+         */
+        beforeEach(function(done){
+            reportServicePage.waitForElement(reportServicePage.tablesListDivEl).then(function(){
+                done();
+            });
+        });
+
+        /**
+        * Test method. Loads the first table containing 10 fields (10 columns). The table report (griddle) width should expand past the browser size
+        * to give all columns enough space to show their data.
+        */
+        it('Table report should expand width past the browser size to show all available data (large num columns)', function(done) {
+            // Select the table
+            reportServicePage.tableLinksElList.get(3).click().then(function(){
+                // Open the reports list
+                reportServicePage.reportHamburgersElList.get(0).click();
+                // Select the report
+                reportServicePage.reportLinksElList.get(0).click();
+                // Make sure the table report has loaded
+                reportServicePage.waitForElement(reportServicePage.loadedContentEl).then(function() {
+                    // Sleep for animation
+                    e2eBase.sleep(2000);
+                    // Check there is a scrollbar in the griddle table
+                    var fetchRecordPromises = [];
+                    fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('scrollWidth'));
+                    fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('clientWidth'));
+                    //When all the dimensions have been fetched, assert the values match expectations
+                    promise.all(fetchRecordPromises).then(function(dimensions) {
+                        expect(Number(dimensions[0])).toBeGreaterThan(Number(dimensions[1]));
+                        reportServicePage.reportsBackLinkEl.click();
                         done();
                     });
                 });
             });
         });
+
         /**
-         * Test method. Loads the first table containing 10 fields (10 columns). The table report (griddle) width should expand past the browser size
-         * to give all columns enough space to show their data.
-         */
-        it('Should expand width past the browser size to show all available data and columns (large data sets)', function(done) {
-            browser.wait(EC.visibilityOf(requestAppsPage.tablesDivEl), 5000).then(function() {
-                // Select the table to load
-                requestAppsPage.tableLinksElList.get(0).click();
-                // Now on the reportServicePage (shows the nav with a list of reports you can load)
-                var reportServicePage = new ReportServicePage();
-                // Wait until the nav has loaded
-                browser.wait(EC.visibilityOf(reportServicePage.navStackedEl), 5000).then(function() {
-                    // Select the report
-                    reportServicePage.navLinksElList.get(1).click();
-                    // Make sure the table report has loaded
-                    browser.wait(EC.visibilityOf(reportServicePage.griddleContainerEl), 5000).then(function() {
-                        // Sleep for animation
-                        e2eBase.sleep(2000);
-                        // Check there is a scrollbar in the griddle table
-                        var fetchRecordPromises = [];
-                        fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('scrollWidth'));
-                        fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('clientWidth'));
-                        //When all the dimensions have been fetched, assert the values match expectations
-                        promise.all(fetchRecordPromises).then(function(dimensions) {
-                            expect(Number(dimensions[0])).toBeGreaterThan(Number(dimensions[1]));
-                            done();
+        * Test method. Loads the second table containing 3 fields (3 columns). The table report (griddle) width should expand
+        * it's columns to fill the available space (and not show a scrollbar).
+        */
+        it('Table report should expand width to take up available space (small num of columns)', function(done) {
+            // Select the table
+            reportServicePage.tableLinksElList.get(4).click().then(function(){
+                // Open the reports list
+                reportServicePage.reportHamburgersElList.get(1).click();
+                // Select the report
+                reportServicePage.reportLinksElList.get(0).click();
+                // Make sure the table report has loaded
+                reportServicePage.waitForElement(reportServicePage.loadedContentEl).then(function() {
+                    // Sleep for animation
+                    e2eBase.sleep(2000);
+                    // Check there is no scrollbar in the griddle table
+                    var fetchRecordPromises = [];
+                    fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('scrollWidth'));
+                    fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('clientWidth'));
+                    //When all the dimensions have been fetched, assert the values match expectations
+                    promise.all(fetchRecordPromises).then(function(dimensions) {
+                        expect(Number(dimensions[0])).not.toBeGreaterThan(Number(dimensions[1]));
+                        reportServicePage.reportsBackLinkEl.click();
+                        done();
+                    });
+                });
+            });
+        });
+
+        /**
+        * Test method. The left hand nav should shrink responsively across the 4 breakpoints as the browser is re-sized
+        */
+        it('Left hand nav should shrink responsively', function(done) {
+            // Select the table
+            reportServicePage.tableLinksElList.get(3).click().then(function(){
+                // Open the reports list
+                reportServicePage.reportHamburgersElList.get(0).click();
+                // Select the report
+                reportServicePage.reportLinksElList.get(0).click();
+                // Make sure the table report has loaded
+                reportServicePage.waitForElement(reportServicePage.loadedContentEl).then(function() {
+                    // Resize browser at different widths to check responsiveness
+                    e2eBase.resizeBrowser(1500, 600).then(function() {
+                        reportServicePage.assertNavProperties('xlarge', true, '399');
+                    }).then(function() {
+                        e2eBase.resizeBrowser(1280, 600).then(function() {
+                            reportServicePage.assertNavProperties('large', true, '299');
+                        }).then(function() {
+                            e2eBase.resizeBrowser(1024, 600).then(function() {
+                                reportServicePage.assertNavProperties('medium', true, '199');
+                            }).then(function() {
+                                e2eBase.resizeBrowser(640, 600).then(function() {
+                                    reportServicePage.assertNavProperties('small', false, '0');
+                                }).then(function() {
+                                    // Reset the browser for other tests
+                                    e2eBase.resizeBrowser(widthTest, heightTest);
+                                    done();
+                                });
+                            });
                         });
                     });
                 });
             });
         });
-        /**
-         * Test method. Loads the second table containing 3 fields (3 columns). The table report (griddle) width should expand
-         * it's columns to fill the available space (and not show a scrollbar).
-         */
-        it('Should expand the width to take up available space in the browser for a small number of columns', function(done) {
-            browser.wait(EC.visibilityOf(requestAppsPage.tablesDivEl), 5000).then(function() {
-                // Select the table to load
-                requestAppsPage.tableLinksElList.get(1).click();
-                // Now on the reportServicePage (shows the nav with a list of reports you can load)
-                // Wait until the nav has loaded
-                var reportServicePage = new ReportServicePage();
-                browser.wait(EC.visibilityOf(reportServicePage.navStackedEl), 5000).then(function() {
-                    // Select the report
-                    reportServicePage.navLinksElList.get(1).click();
-                    // Make sure the table report has loaded
-                    browser.wait(EC.visibilityOf(reportServicePage.griddleContainerEl), 5000).then(function() {
-                        // Sleep for animation
-                        e2eBase.sleep(2000);
-                        // Check there is no scrollbar in the griddle table
-                        var fetchRecordPromises = [];
-                        fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('scrollWidth'));
-                        fetchRecordPromises.push(reportServicePage.loadedContentEl.getAttribute('clientWidth'));
-                        //When all the dimensions have been fetched, assert the values match expectations
-                        promise.all(fetchRecordPromises).then(function(dimensions) {
-                            expect(Number(dimensions[0])).not.toBeGreaterThan(Number(dimensions[1]));
-                            done();
-                        });
-                    });
-                });
-            });
-        });
-        ///**
-        // * Return to the table selection page by clicking on the Home link in the left nav
-        // */
-        afterEach(function(done) {
-            var reportServicePage = new ReportServicePage();
-            browser.wait(EC.visibilityOf(reportServicePage.navStackedEl), 5000);
-            reportServicePage.appsHomeLinkEl.click();
-            done();
-        });
+
         /**
          * After all tests are done, run the cleanup function in the base class
          */
