@@ -1,6 +1,7 @@
 // Karma configuration
 //
 var path = require("path");
+var webpack = require('webpack');
 var nodeModulesPath = path.resolve(__dirname, "node_modules");
 
 module.exports = function(config) {
@@ -38,7 +39,10 @@ module.exports = function(config) {
                             path.resolve(__dirname, "client-react/test")
                         ],
                         exclude: [nodeModulesPath],
-                        loader: "babel-loader?plugins=babel-plugin-rewire"
+                        loader: "babel-loader",
+                        query: {
+                            plugins: ['babel-plugin-rewire', 'babel-plugin-rewire-ignore-coverage']
+                        }
                     },
                     {
                         // all css files can be required into js files with this
@@ -66,6 +70,10 @@ module.exports = function(config) {
                         include: [
                             path.resolve(__dirname, "client-react/src")
                         ],
+                    },
+                    {
+                        test   : /\.woff|\.woff2|\.svg|.eot|\.ttf/,
+                        loader : 'url?prefix=font/&limit=10000'
                     }
                 ],
                 postLoaders: [
@@ -82,6 +90,16 @@ module.exports = function(config) {
                     }
                 ]
             },
+            plugins: [
+                // Define the build run-time environment.  Configure to run as a PROD build.
+                // This will allow us to write unit tests against the prod configuration.
+                // ONE and ONLY ONE variable is to be set to true.
+                new webpack.DefinePlugin({
+                    __QB_PROD__: JSON.stringify(true),
+                    __QB_TEST__: JSON.stringify(false),
+                    __QB_LOCAL__: JSON.stringify(false)
+                })
+            ],
             watch: true
         },
         webpackServer: {
