@@ -1,25 +1,22 @@
 'use strict';
 
+var assert = require('assert');
 var app = require('../../app');
-var request = require('supertest');
+var request = require('request');
 var config = require('../../config/environment');
 var apiBase = require('./api.base.js')(config);
-require('should');
+var testConsts = require('./api.test.constants');
 
 describe('GET /api/health', function() {
     it('should respond with JSON object with sqlDate', function(done) {
-        request(app)
-                .get(apiBase.resolveHealthEndpoint())
-                .set('Accept', 'application/json')
-                .expect(200)
-                .expect('Content-Type', /json/)
-                .end(function(err, res) {
-                    if (err) {
-                        return done(err);
-                    }
-                    res.body.should.be.instanceof(Object);
-                    res.body.should.have.keys('sqlDate');
-                    done();
-                });
+        this.timeout(testConsts.INTEGRATION_TIMEOUT);
+        var opts = {
+            url   : config.DOMAIN + apiBase.resolveHealthEndpoint(),
+            method: 'GET'
+        };
+        request(opts, function(error, response) {
+            assert.equal(response.statusCode, 200, 'Unexpected status code.');
+            done();
+        });
     });
 });
