@@ -9,52 +9,22 @@
     module.exports[403] = function unauthorized(req, res) {
         var viewFilePath = '403.html';
         var statusCode = 403;
-        var result = {
-            status: statusCode
-        };
-
-        res.status(result.status);
-        //respond with json if requested - for swagger ajax
-        if (req.headers.accept === consts.APPLICATION_JSON) {
-            res.json(result, result.status);
-        } else {
-            res.render(viewFilePath, function(err) {
-                if (err) {
-                    return res.json(result, result.status);
-                }
-                res.render(viewFilePath);
-            });
-        }
-        log.logRequest(req, __filename);
+        processError(req, res, viewFilePath, statusCode);
     };
 
     module.exports[404] = function pageNotFound(req, res) {
         var viewFilePath = '404.html';
         var statusCode = 404;
-        var result = {
-            status: statusCode
-        };
-
-        res.status(result.status);
-
-        //respond with json if requested - for swagger ajax
-        if (req.headers.accept === consts.APPLICATION_JSON) {
-            res.json(result, result.status);
-        } else {
-            res.render(viewFilePath, function(err) {
-                if (err) {
-                    return res.json(result, result.status);
-                }
-                res.render(viewFilePath);
-            });
-        }
-        log.logRequest(req, __filename);
-
+        processError(req, res, viewFilePath, statusCode);
     };
 
     module.exports[500] = function internalServerError(req, res) {
         var viewFilePath = '500.html';
         var statusCode = 500;
+        processError(req, res, viewFilePath, statusCode);
+    };
+
+    function processError(req, res, viewFilePath, statusCode) {
         var result = {
             status: statusCode
         };
@@ -71,7 +41,7 @@
                 res.render(viewFilePath);
             });
         }
-        log.logRequest(req, __filename);
-    };
+        log.error({req: req}, 'process error');
+    }
 
 }());
