@@ -3,124 +3,262 @@
 
 var request = require('supertest');
 var should = require('should');
+var assert = require('assert');
+var rewire = require('rewire');
+
+var routeGroups = require('../../routes/routeGroups');
+var envConsts = require('../environment/environmentConstants');
+
 describe('Request test always use ssh', function() {
 
-    var server = require('express');
-    var app = server();
+    it('test undefined config environment should throw an exception', function() {
 
-    var config = require('../expressConfig')(app);
+        var exception = false;
+        var mockConfig = {
+            //  no environment defined
+            hasSslOptions: function() {
+                return false;
+            }
+        };
 
-    // ensure ssl values are set be able to run secure http test
-    it('config should have sslPort ', function() {
-        should.exist(config.sslPort);
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+
+        try {
+            testExpress(testapp);
+        } catch (e) {
+            exception = true;
+        }
+
+        assert(exception);
     });
 
-    it('sslPort should be numeric', function() {
-        config.sslPort.should.be.an.instanceOf(Number);
+    it('test undefined route group  and IP is set to their defaults', function() {
+
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.TEST,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+            //  no route group defined
+            //  no ip or DOMAIN
+        };
+
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert.equal(newConfig.routeGroup, routeGroups.DEFAULT);
+        assert.equal(newConfig.ip, 'localhost');
+    });
+});
+
+describe('Test environments are setup correctly', function() {
+
+    it('test production environment configuration', function() {
+
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.PRODUCTION,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+            //  no route group defined
+        };
+
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert(newConfig.isProduction);
     });
 
-    it('should have NODE_TLS_REJECT_UNAUTHORIZED defined', function() {
-        //environment var to allow self signed cert
-        should.exist(process.env.NODE_TLS_REJECT_UNAUTHORIZED);
+    it('test environment configuration', function() {
+
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.TEST,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+        };
+
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert(!newConfig.isProduction);
     });
 
+    it('PRE-PROD environment configuration', function() {
 
-    it('NODE_TLS_REJECT_UNAUTHORIZED should =0 to support selfsigned cert', function() {
-        //environment var to allow self signed cert
-        process.env.NODE_TLS_REJECT_UNAUTHORIZED.should.equal('0');
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.PRE_PROD,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+        };
+
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert(newConfig.isProduction);
     });
 
+    it('DEVELOPMENT environment configuration', function() {
 
-    describe('hasSslOptions method', function() {
-        /* eslint no-unused-expressions:0 */
-        it('should correctly check for ssl config settings', function(done) {
-            var origEnvVal = process.env.NODE_ENV;
-            process.env.NODE_ENV = 'dummy';
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.DEVELOPMENT,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+        };
 
-            // init the app server with dummy env and test the configs ssl function
-            var testapp = require('express')();
-            var testconfig = require('../expressConfig')(testapp);
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
 
-            should.exist(testconfig.hasSslOptions);
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
 
-            describe('incomplete sslkey object', function() {
-                it('hasSslOptions should return false ', function() {
-                    testconfig.SSL_KEY = null;
-                    testconfig.hasSslOptions().should.be.false;
+        assert(!newConfig.isProduction);
+    });
 
-                    testconfig.SSL_KEY = {};
-                    testconfig.hasSslOptions().should.be.false;
+    it('INTEGRATION environment configuration', function() {
 
-                    testconfig.SSL_KEY = {private: null};
-                    testconfig.hasSslOptions().should.be.false;
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.INTEGRATION,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT'
+        };
 
-                    testconfig.SSL_KEY = {cert: null};
-                    testconfig.hasSslOptions().should.be.false;
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
 
-                    testconfig.SSL_KEY = {private: null, cert: null};
-                    testconfig.hasSslOptions().should.be.false;
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
 
-                    testconfig.SSL_KEY = {private: '', cert: ''};
-                    testconfig.hasSslOptions().should.be.false;
+        assert(!newConfig.isProduction);
+    });
 
-                    testconfig.SSL_KEY = {private: 'p', cert: ''};
-                    testconfig.hasSslOptions().should.be.false;
+    it('test domain configuration', function() {
 
-                    testconfig.SSL_KEY = {private: '', cert: 'c'};
-                    testconfig.hasSslOptions().should.be.false;
-                });
-            });
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.TEST,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT',
+            DOMAIN: 'http://domain.com'
+        };
 
-            describe('valid sslkey object', function() {
-                it('hasSslOptions should return true ', function() {
-                    testconfig.SSL_KEY = {private: 'thing1', cert: 'thing2'};
-                    testconfig.hasSslOptions().should.be.true;
-                });
-            });
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
 
-            process.env.NODE_ENV = origEnvVal;
-            done();
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert.equal(newConfig.ip, 'domain.com');
+    });
+
+    it('test ip configuration', function() {
+
+        var exception = false;
+        var mockConfig = {
+            env: envConsts.TEST,
+            hasSslOptions: function() {
+                return false;
+            },
+            root: 'testRoot',
+            client: 'REACT',
+            ip: '1.2.3.4'
+        };
+
+        var testExpress = rewire('../expressConfig');
+        testExpress.__set__({'config': mockConfig});
+
+        var testapp = require('express')();
+        var newConfig = testExpress(testapp);
+
+        assert.equal(newConfig.ip, mockConfig.ip);
+    });
+});
+
+
+describe('hasSslOptions method', function() {
+    /* eslint no-unused-expressions:0 */
+
+    // init the app server with dummy env and test the configs ssl function
+    var testapp = require('express')();
+    var testconfig = require('../expressConfig')(testapp);
+
+    it('should correctly check for ssl config settings', function() {
+        var origEnvVal = process.env.NODE_ENV;
+        process.env.NODE_ENV = 'dummy';
+        should.exist(testconfig.hasSslOptions);
+        process.env.NODE_ENV = origEnvVal;
+    });
+
+    it('hasSslOptions should return false ', function() {
+        testconfig.SSL_KEY = null;
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {private: null};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {cert: null};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {private: null, cert: null};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {private: '', cert: ''};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {private: 'p', cert: ''};
+        testconfig.hasSslOptions().should.be.false;
+
+        testconfig.SSL_KEY = {private: '', cert: 'c'};
+        testconfig.hasSslOptions().should.be.false;
+    });
+
+    it('valid sslkey object', function() {
+        it('hasSslOptions should return true ', function() {
+            testconfig.SSL_KEY = {private: 'thing1', cert: 'thing2'};
+            testconfig.hasSslOptions().should.be.true;
         });
-    });
-
-    // ensure request use secure http
-    it('server http request should return redirect to https with sslPort', function(done) {
-        if (config.hasSslOptions()) {
-            request('http://' + config.ip + ':' + config.port)
-                    .get('/')
-                    .expect(302)
-                    .expect(function(res) {
-                        var foundSslPort = (res.headers.location.indexOf(config.sslPort + '/') > -1);
-                                //expect returns true on fail (foundSslPort is expected)
-                        return !foundSslPort;
-                    })
-                    .end(function(err) {
-                        done(err);
-                    });
-        } else {
-            done();
-        }
-    });
-
-    // ensure request use secure http if specified
-    it('server should support https', function(done) {
-        //test ssl cert if available
-        if (config.hasSslOptions()) {
-            should.exist(process.env.NODE_TLS_REJECT_UNAUTHORIZED);
-            request('https://' + config.ip + ':' + config.sslPort)
-                    .get('/')
-                    .expect(200)
-                    .expect(function(res) {
-                                //expect returns true on fail (no redirect is expected)
-                        return true === res.redirect;
-                    })
-                    .end(function(err) {
-                        done(err);
-                    });
-        } else {
-            done();
-        }
-
     });
 });
 
