@@ -10,22 +10,21 @@
 
     module.exports = function(config) {
         if (!config.isProduction && !config.noHotLoad) {
+
             var webpack = require('webpack');
             var WebpackDevServer = require('webpack-dev-server');
-            var hotPort = config.webpackDevServerPort || 3000;
             var webpackConfig = require('../webpack.config.js');
+            var hotPort = config.webpackDevServerPort || 3000;
 
             webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
             webpackConfig.entry = [
                     // WebpackDevServer host and port
                 'webpack-dev-server/client?http://0.0.0.0:' + hotPort,
-                    //'webpack-dev-server/client?http://localhost:' + hotPort,
                     // 'only-dev-server' prevents reload on syntax errors
                 'webpack/hot/only-dev-server'
             ].concat(webpackConfig.entry);
 
             webpackConfig.output.publicPath = 'http://' + config.ip + ':' + hotPort + webpackConfig.output.publicPath;
-            //webpackConfig.output.publicPath = '/dist/';
 
             var compiler = webpack(webpackConfig);
             // we start a webpack-dev-server with our config
@@ -37,9 +36,7 @@
                 // webpack-dev workflow, but we want the 'in-memory' files to be fetched from
                 // this path  e.g http://localhost:3000/dist/
                 publicPath : '/dist/',
-                //publicPath        : 'http://localhost:' + hotPort + webpackConfig.output.publicPath,
 
-                //contentBase        : webpackConfig.contentBase,
                 // hot -  Enable special support for Hot Module Replacement (HMR)
                 // Page is not updated, but a 'webpackHotUpdate' message is send to the content
                 // Use 'webpack/hot/dev-server' as additional module in your entry point
@@ -51,8 +48,6 @@
                 inline             : true,
 
                 progress: true,
-
-                //debug: true,
 
                 // historyApiFallback -  to access dev server from arbitrary url.
                 historyApiFallback: true,
@@ -75,7 +70,9 @@
             log.info('webpackConfig Settings :' + JSON.stringify(webpackConfig));
             log.info('devServerConfig Settings :' + JSON.stringify(devServerConfig));
 
-            var hotServer = new WebpackDevServer(compiler, devServerConfig);
+            var hotServer =  config.hotServer ?
+                new config.hotServer(compiler, devServerConfig) :
+                new WebpackDevServer(compiler, devServerConfig);
             log.info('Hot webpack-dev-server Settings :' + JSON.stringify(hotServer));
             hotServer.listen(hotPort, config.ip, function(err) {
                 if (err) {
@@ -83,8 +80,6 @@
                 }
                 log.info('Hot webpack-dev-server Listening at ' + config.ip + ' port:' + hotPort);
             });
-            //console.log('\nwebpackConfig Settings :' + JSON.stringify(webpackConfig, null, 2));
-
         }
 
     };
